Dataset and preprocessing
Added the dataset to the repo and completed preprocessing. Code is commented to explain as well as the there is an explanation of what the arrays are at the end.

Other preprocessing we can do is noise reduction feature scaling, histogram equalization. I don't think we have to add these unless we start running into issues