GCN failed but alternative
Could not get GCNN working had to switch to CNN with graph based features also have an FNN as well.

posting this for historics. Next push will be cleaned up and back to one file. Next is tuning again for these two and will be completed. DNN is completed already as well.